Rewrite of the overriding DNS router
Removes the need to use a "fallback" address such as `8.8.8.8` in favor
of instead utilizing the original router address. This is done by
first establishing the connection to the original router that can be
used for fallbacks, then adding an extra firewall rule that prevents
rerouting of requests that stems from the local address of that
connection.

Since this makes the current `--fallback` completely redundant, that
option was removed too.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>